Remove kharris@nexus-tech.net from MAINTAINERS
Mail to kharris@nexus-tech.net bounces because the user doesn't exist
anymore. You can't be a maintainer without a valid e-mail address, so
move all boards that used to be maintained by Kyle Harris to the
"orphaned" list.

Currently, only PowerPC has a list of orphaned boards, so this patch
creates one for ARM as well.

Signed-off-by: Haavard Skinnemoen <haavard.skinnemoen@atmel.com>