ddr cfg: DRAM_RESET needs 0x00020030
The old value of 0x000e0030 will cause ethernet to
not work under u-boot for 1 SOM board. It will also
cause droppped packets for boards using KSZ9021.
I have no explanation as to why.

But this is a correct change, the TRM will be updated
to show that 00b is the only valid setting for bits
19-18 of DRAM_RESET.

Signed-off-by: Troy Kisky <troy.kisky@boundarydevices.com>