nitrogen6x: config: configure usb_ether
Provide fixed USB networking mac addresses for host and client to enable
static configuration of host network stacks.

Include a command 'usbrecover' both to illustrate the use of the USB
ethernet gadget and also to allow quick booting of a kernel (uImage)
and ram disk (uramdisk).

Details and commentary are available here:
        http://boundarydevices.com/u-boot-2014-01/#usbrecover

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>