nitrogen6x: Allow override of board-type detection.
The automatic detection mechanism for determining whether
a board is a SABRE Lite or Nitrogen6x is to sample the WL_IRQ
pin when the wl1271 chip (if connected) is in reset.

See this post for details:
http://boundarydevices.com/differences-sabre-lite-nitrogen6x

Unfortunately, this requires an additional components to
be populated.

This patch allows override through an environment variable
'forcen6', which allows a reduced BOM on boards which won't
have the Wi-Fi option installed.

        U-Boot > setenv forcen6 1
        U-Boot > saveenv && reset

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>