nit6xlite: boot scripts: add flag "sdphys" to choose proper SD card root
Kernel versions 3.10.31+ choose SD card device names (/dev/mmcblkN)
based on their physical path rather than order of initialization.

Add support for a flag 'sdphys' to choose this when adding the 'root='
clause to the kernel command line.

In order to boot an SD card with a newer kernel, set 'sdphys' to
anything non-blank and save:
        U-Boot > setenv sdphys 1
        U-Boot > saveenv

To clear this out when booting an older kernel, clear the
environment variable:
        U-Boot > setenv sdphys
        U-Boot > saveenv

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>