nitrogen6_max: update to compile with MX6QDL
Because the board file for Nitrogen6_max is configured for
use with SPL (both MX6Q and MX6DL registers), the MX6QDL
flag must be used for compilation of all C sources.

In particular, arch/arm/imx-common/iomux-v3.c needs this
to interpret the register lists.

Because we're still initializing RAM in the DCD, the
configuration file nitrogen6_max_defconfig must be compiled
with CONFIG_MX6Q and not CONFIG_MX6QDL.

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>