nitrogen6x: Add bootscript for 3.10.53+ on Yocto
Support for the "video=" clauses is in the process of being deprecated
and specifically doesn't work on 3.10.53 for LVDS displays (ldb driver).

This bootscript inverts the logic and rather than adding a video= clause
for each display found, it removes the device-tree branch for display
that are **not found**.

It also fully requires device tree, where the previous version
(6x_bootscript-yocto.txt) tried to guess whether a device-tree-enabled
kernel was present on the boot partition.

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>