ls: rename MIC detect to HP_DET, add HP_DET2
HP_DET (EIM_A16) is shown as not connected in the current
schematic. Include it as HP_DET2 in case the layout doesn't
match.

MIC_DET (EIM_A16) is a misnomer, since all we can tell is
whether a headset is plugged in. Detecting if the headset has
a microphone requires impedance checking on the signal lines.
Rename this to HP_DET.

Signed-off-by: Eric Nelson <eric.nelson@boundarydevices.com>