libfdt: Add overlay application function
PD#SWPL-772

The device tree overlays are a good way to deal with user-modifyable
boards or boards with some kind of an expansion mechanism where we can
easily plug new board in (like the BBB, the Raspberry Pi or the CHIP).

Add a new function to merge overlays with a base device tree.

Change-Id: I3f96e908248a0734979875874ae31e109f1d3d0f
Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>
Signed-off-by: jiamin ma <jiamin.ma@amlogic.com>