6x_bootscript-mainline: change default bootdir to be /
Since this script was made for Buildroot at first, it was easier to
have bootdir defaulted to /boot/ since this is where the components
are located.

But Otavio recently added Yocto mainline support [1] where all the
components are in / so it is actually better to have the same behavior
as other Yocto 6x_bootscript.

Note that for Buildroot support, the 6x_bootscript is part of the
source code [2] so it will be updated for mainline support in a near
future.

[1] https://lists.yoctoproject.org/pipermail/meta-freescale/2017-March/020290.html
[2] https://git.buildroot.net/buildroot/tree/board/boundarydevices/common/6x_bootscript.txt

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>