net: eth_legacy: fix on_ethaddr possible loop
There's a corner case that ends up in an infinite loop currently.
If you set ethaddr to 00:00:00:00:00:00 in U-Boot, it works ok as there
is an env callback that ensures the value is correct.
However, if one uses fw_setenv utility from user-space to set ethaddr to
a wrong value, then the first call to eth_write_hwaddr() will end up in
a recursive loop as it always reads the ethaddr from env instead of
using the new value about to be applied.

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>