configs: mt8365_pumpkin: set CONFIG_LMB_MAX_REGIONS to 16
We found u-boot fails to boot when adding new reserved regions in
device tree:

===
ERROR: Failed to allocate 0x1f617 bytes below 0xc0000000.
device tree - allocation error
FDT creation failed!
===

The reason is that the u-boot memory allocator (LMB) has a default
limitation on the max number of reserved regions (8), and adding
new region exceeds the maximum number, which causes the failure.

Hence we increse the number from 8 to 16 for fixing the issue.

Change-Id: I5fe0b8ce6d746fd2c0a4c7cc0437ee28a8963273
Signed-off-by: Ramax Lo <ramax.lo@mediatek.com>
Signed-off-by: Macpaul Lin <macpaul.lin@mediatek.com>