gpio : pca953x_gpio : add support for PCAL6534
Ported from the Linux driver changes by
Martyn Welch

https://github.com/nxp-imx/linux-imx.git
13c5d4ce806026a3d06efa2fcc1e5a0cb875df29

This device is broadly a 34-bit version
of the PCAL6524. However, whilst the registers are broadly what you'd
expect for a 34-bit version of the PCAL6524, the spacing of the registers
has been compacted. This has the unfortunate effect of breaking the bit
shift based mechanism that is employed to work out register locations used
by the other chips supported by this driver. To accommodate this, a callback
function has been added to allow alterate implementations of
pca953x_recalc_addr() for the PCAL6534.

Signed-off-by: Simon Gaynor <simon.gaynor@lairdconnect.com>