GENIO: genio-510-evk: enable dsi bootlogo drivers
Since the Genio-510-EVK uses the same chip as the genio-700-evk,
the DSI drivers can be reused.

The issue with disp_pwm has also been resolved in this change.
Initially, disp_pwm_mt8188.c used disp_pwm0, but it was found
that disp_pwm1 is the correct PWM for the DSI panel. Therefore,
disp_pwm1 should be used instead.

On the other hand, since neither the genio-700-evk nor the
genio-510-evk have a clock driver in U-Boot, register control
should still be used to enable the PWM clocks.

Change-Id: I1e29d72791eb73c60bb9cb4f19143157af4ee9b7
Signed-off-by: Tommy Chen <tommyyl.chen@mediatek.com>