GENIO: mt8370: add bootlogo configs
a. Set splashsource as mmc_fs to read the logo file from an mmc_fs
    of an image partition.
b. Set splashfile as logo.bmp to assign the logo file name.
c. Set splashdevpart as bootassets to read the logo from the
    bootassets partition.
d. Set splashpos as m,m to position the logo in the middle of the
    panel.

Change-Id: I1e86d1ae75fbcffa8e9c6233fcb3be4dab4fd6c0
Signed-off-by: Tommy Chen <tommyyl.chen@mediatek.com>