LF-10422-1 rtc: Fix date reset fail
The date reset cmd will reset the rtc of seq 0. But the rtc of seq 0
is changed from rtc@53 into rtc_emul when the fan peng patch enable
the emul rtc.

This patch assign the rtc0 to rtc@53. In addition, the rtc@53 pcf2131
driver probe will check this rtc if it exists because there is no pcf2131
on imx93 board before B1.

Signed-off-by: Joy Zou <joy.zou@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>