arm: mediatek: spl: support normal and download mode
The U-Boot SPL is used for two different modes:
- Normal mode:
The device is power-on, BL1 will look at the BOOT0 in the eMMC.
It founds the U-Boot SPL and we load all the next ATF images.
This bootflow is supported by spl_atf.c

- Download mode:
The device is power-on with a key combo which trigger the Download
mode.
In this mode the BL1 listen on USB and wait a Download Agent (DA).
In our case the Dowload Agent is U-Boot SPL.
The U-Boot SPL is loaded and goes in fastboot mode.
This bootflow is supported by spl_fastboot.c

Notes:
When we flash the first time a blank device, it may happen that the
boot partition in eMMC is not enabled.
We must setup the eMMC in order to boot from a partition.
This is done by spl_fastboot.c

Signed-off-by: Julien Masson <jmasson@baylibre.com>