LFU-638 imx95-Titan: Fix I2C address conflict and I2C7 speed
The I2C address of typec controller is 0x50 which conflicts with
address of EEPROM on Verdin DSI to HDMI adapter. Once the adapter
is installed, typec can't work.

Toradex confirmed to remove R97 on carrier board to change typec
address to 0x52 to fix conflict. So update the address in board
codes and DTS.

Also the Verdin board does not support 1Mhz I2C speed. Have to
change the I2C7 speed to 400Khz in DTS

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>