[WORKAROUND] drivers: usb: mtu3: add delay before soft disconnect
D+/D- lines are used by the usb framework in order to do a software
controlled usb_gadget_disconnect().

Currently, when this disconnection occurs due to a fastboot reboot
command, the lines are disabled to soon, and fastboot command returns
an error. Add a delay so that the gagdet disconnection occurs at
the goot timing.

Signed-off-by: Vitor Sato Eschholz <vsatoes@baylibre.com>