bootstd: android: allow reading whole partition
Commit 30a35f2259 (bootstd: android: don't read whole partition
sizes) only took into account the reading of partitions which the
size can be discovered by parsing the boot header.

One might also want to read partitions that we don't know the size.
In that case, the whole partition should be read.

Allow read_slotted_partition to read the whole partition when the
size is not known beforehand. In that case, image_size argument
should be zero.

Signed-off-by: Vitor Sato Eschholz <vsatoes@baylibre.com>